- 调整代码

+++ Engine/Renderer/GLES3/src/GLES3FrameBuffer.cpp
@@ -42,7 +42,7 @@ int CGLES3FrameBuffer::GetHeight(void) const
 bool CGLES3FrameBuffer::SetAttachmentTexture(int indexAttachment, CGfxRenderTexturePtr &ptrAttachmentTexture)
 {
 	if (ptrAttachmentTexture->GetWidth() == m_width && ptrAttachmentTexture->GetHeight() == m_height &&
-		indexAttachment >= 0 && indexAttachment < m_ptrAttachmentTextures.size()) {
+		indexAttachment >= 0 && indexAttachment < (int)m_ptrAttachmentTextures.size()) {
 		m_ptrAttachmentTextures[indexAttachment] = ptrAttachmentTexture;
 		return true;
 	}
@@ -53,7 +53,7 @@ bool CGLES3FrameBuffer::SetAttachmentTexture(int indexAttachment, CGfxRenderText
 
 CGfxRenderTexturePtr CGLES3FrameBuffer::GetAttachmentTexture(int indexAttachment) const
 {
-	if (indexAttachment >= 0 && indexAttachment < m_ptrAttachmentTextures.size()) {
+	if (indexAttachment >= 0 && indexAttachment < (int)m_ptrAttachmentTextures.size()) {
 		return m_ptrAttachmentTextures[indexAttachment];
 	}
 	else {
